2023-3-13 17:11

+++ src/views/MyDocument/index.vue
@@ -6,40 +6,52 @@
       <span class="my-main-doc-menu-title" @click="showProjectDiv" :class="{ active: currentIndex === 1 }">任务管理</span>
     </div>
     <div class="my-main-doc-toggle">
-<UploadFileBtn className="my-main-doc-toggle-upload" v-show="showUpload" @btnClick="getData" />
+      <div class="my-main-doc-toggle-upload-header" v-show="showUpload">
+        <el-select v-model="curValue" placeholder="当前语言" filterable>
+          <el-option v-for="item in data.Lang " :key="item.value" :label="item.label" :value="item.value"></el-option>
+        </el-select>
+        <img class="my-main-doc-toggle-upload-header-trans" src="@/assets/myText/translate.svg">
+        <el-select v-model="tarValue" placeholder="目前只支持中文">
+          <el-option v-for="item in data.lang " :key="item.value" :label="item.label" :value="item.value"></el-option>
+        </el-select>
+      </div>
+      <UploadFileBtn className="my-main-doc-toggle-upload" :fromLang="this.data.Lang[this.curValue-1].name" v-show="showUpload" @btnClick="getData" />
       <div class="my-main-doc-toggle-project" v-show="showProject">
         <div class="my-main-doc-toggle-project-header">
           <div class="my-main-doc-toggle-project-header-left">
-            <el-input placeholder="请输入内容" v-model="inputSearch" style="width: 200px"></el-input>
+            <el-input placeholder="请输入文件名" v-model="inputSearch" style="width: 200px"></el-input>
             <el-button type="primary" style="background-color: #3867FF; margin-left: 10px" @click="getDataByid">查询</el-button>
           </div>
           <div class="my-main-doc-toggle-project-header-right">
             <!-- <el-select v-model="transValue" placeholder="选择翻译状态" style=" margin-right: 10px">
               <el-option v-for="item in data.transState" :key="item.value" :label="item.label" :value="item.value"></el-option>
             </el-select> -->
-            <el-button type="primary" style="background-color: #3867FF; margin-left: 10px" @click="getData">重置</el-button>
+            <el-button type="primary" style="background-color: #3867FF; margin-left: 10px" @click="getData">刷新</el-button>
           </div>
         </div>
         <div class="my-main-doc-toggle-project-table">
           <el-table :data="data.workList" border style="width: 100%" max-height="580px"> 
-          <el-table-column prop="" align="center" label="选中" style="width: 11%"><el-checkbox></el-checkbox></el-table-column>
-          <el-table-column prop="id" label="任务ID" style="width: 11%"></el-table-column>
-          <el-table-column prop="filename" label="文件名" style="width: 11%"></el-table-column>
-          <el-table-column prop="size" label="文件大小" style="width: 11%"></el-table-column>
-          <el-table-column prop="" label="目标语言" style="width: 11%">中文</el-table-column>
-          <el-table-column prop="status" label="翻译状态" style="width: 11%">
+          <el-table-column prop="index" align="center" label="序号" width="40%"></el-table-column>
+          <el-table-column prop="id" label="任务ID" style="width: 10%"></el-table-column>
+          <el-table-column prop="filename" label="文件名" style="width: 10%"></el-table-column>
+          <el-table-column prop="size" label="文件大小" style="width: 10%"></el-table-column>
+          <el-table-column prop="from" label="当前语言" style="width: 10%">
+          </el-table-column>
+          <el-table-column prop="" label="目标语言" width="100%">中文</el-table-column>
+          <el-table-column prop="status" label="翻译状态" style="width: 10%">
             <template slot-scope="{ row }">
               {{ mapStatus[row.status] }}
             </template>
           </el-table-column>
-          <el-table-column prop="createTime" label="创建时间" style="width: 11%"></el-table-column>
-          <el-table-column prop="transTime" label="开始翻译时间" style="width: 11%"></el-table-column>
-          <el-table-column prop="name" label="操作" style="width: 11%">
+          <el-table-column prop="createTime" label="创建时间" style="width: 10%"></el-table-column>
+          <el-table-column prop="transTime" label="开始翻译时间" style="width: 10%"></el-table-column>
+          <el-table-column prop=""  label="操作" style="width: 15%">
             <template slot-scope="{ row }">
-            <el-button type="danger" @click="deleteData(row)" size="mini">
-              <i class="el-icon-delete"></i> 删除
-            </el-button>
-          </template>
+              <div class="button-container">
+                <el-button type="danger"  @click="deleteData(row)" size="mini">删除</el-button>
+                <el-button type="primary" width="20%" size="mini" style="background-color: #3867FF; margin-left: 10px" @click="getDownload(row)">下载</el-button>
+              </div> 
+            </template>
           </el-table-column>
         </el-table>
         </div>
@@ -49,8 +61,9 @@
 </template>
 <script>
   import UploadFileBtn from '@/components/UploadFileBtn'
-  import data from '@/models/MyDocument/index'
+  import data from '@/models/MyTextnDoc/index'
   import request from '@/utils/request'
+  import { MessageBox  } from 'element-ui'
 export default {
   components:{
     UploadFileBtn
@@ -60,6 +73,8 @@ export default {
       ...data,
       checked: false,
       currentIndex: 0,
+      curValue:4,
+      tarValue:'',
       showUpload: true,
       showProject: false,
       orderValue: '',
@@ -72,15 +87,17 @@ export default {
   this.getData()
   },
   methods:{
+    getDownload(row) {
+      let a = document.createElement('a')
+      a.href = `http://asia.52asia.asia:4040/trans/down?appid=123456&id=${row.id}&lang=zh`
+      a.click()
+    },
     async getDataByid(){
-      console.log('出发')
-      console.log('1',this.data.workList)
       const targetList=this.data.allList.map(item => {if(item.filename === this.inputSearch) return item})
       this.data.workList=[]
-      console.log(targetList)
+      let index=1;
       targetList.map(async item => {
       if(item){
-        console.log(item.id)
       const res= await request({
         url:'/trans/doc',
         method: 'get',
@@ -90,27 +107,51 @@ export default {
           id:item.id
         }
       })
-      console.log(res)
-      this.data.workList=[...this.data.workList,...res.result]
-      } 
-      })
-  },
+      const obj =res.result[0]
+      if( obj.status=== 9 ) obj.status= 4;
+      obj.index=index; index++;
+      let findObj =this.data.Lang.find( objj => objj.name === obj.from)
+      obj.from=findObj.label;
+      findObj =this.data.Lang.find( objj => objj.name === obj.to )
+      obj.to =findObj.label;
+      this.data.workList = [...this.data.workList,obj ]
+      } })
+    },
     async getData(){
+      console.log()
       const res = await request({
         url: '/trans/doc',
         method:'get',
         params:{
           appid: '123456',
-          page: 0
+          page: 0,
+          from: this.data.Lang[this.curValue-1].name,
+          to: "zh_CN"
         }
       })
       console.log(res)
       this.data.allList = res.result
-      this.data.allList.forEach((obj,index)=> { obj.page=res.page; if( obj.status=== 9 ) obj.status= 4; obj.index=index})
+      this.data.allList.forEach( (obj,index) => { 
+        obj.page=res.page; 
+        if( obj.status=== 9 )
+        obj.status= 4;
+        obj.index=index+1;
+        let findObj =this.data.Lang.find( objj => objj.name === obj.from)
+        console.log('123',findObj)
+        obj.from =findObj.label
+        findObj =this.data.Lang.find( objj => objj.name === obj.to)
+        obj.to =findObj.label
+      })
       this.data.workList = this.data.allList
+      console.log('111',this.data.workList)
     },
-    async deleteData(row) {
-      const res = await request({
+    deleteData(row) {
+      MessageBox.confirm('确认删除该数据吗？','提示',{
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'Warning'
+      }).then(async ()=> {
+        const res = await request({
         url:'/trans/doc',
         method:'delete',
         params:{
@@ -120,6 +161,8 @@ export default {
       })
       console.log('删除记录',res)
       this.getData()
+      }).catch(()=>{})
+    
     },
     showUploadDiv() {
       this.showUpload = true;
@@ -155,7 +198,8 @@ export default {
   &-space {
     margin: 0 8px; } }
   &-toggle {
-    margin-top: 40px;
+    margin-top: 20px;
+  
     &-upload{
       display: flex;
       align-items: center;
@@ -164,6 +208,13 @@ export default {
       width:100%;
       height:350px;
       border: 1px dotted #999;
+      &-header{
+        margin-top:60px;
+        margin-bottom:30px;
+        >img{
+          margin: 0 10px;
+        }
+      }
       &-btn {
         width: 220px;
         height: 40px;
@@ -191,5 +242,9 @@ export default {
 }
 .active {
   color: #3867FF;
+}
+
+.button-container {
+  display: flex;
 }
 </style>
